% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppServer.R
\name{AppServer}
\alias{AppServer}
\title{protein'=rep(prot_names[i],length(times)),
z_score'=z_score,}
\usage{
AppServer(input, output, session)
}
\description{
protein'=rep(prot_names[i],length(times)),
z_score'=z_score,
}
